pyproject.toml
[tool.poetry]
name = "ds-util"
version = "0.1.0"
description = "Util scripts for data science projects and Kaggle."
authors = ["kiccho1101 <youodf11khp@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.2.4"
numpy = "^1.20.3"
psutil = "^5.8.0"
torch = "^1.8.1"
mecab-python3 = "^1.0.3"
scikit-learn = "^0.24.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.6b0"
flake8 = "^3.9.2"
isort = "^5.8.0"
pytest-mock = "^3.6.1"
pytest-cov = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
